relocate ListUtility to utils package and reorder modifiers to convention
