+++ src/test/java/commandline/utils/ListUtilityTest.java
@@ -1,4 +1,4 @@
-package commandline.view;
+package commandline.utils;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.ArrayList;
